UnifiedPush support PR by @Golbinex (#306)
* [TF][PUSH] implemented a simple UnifiedPush support

This simple implementation only triggers the networks to resume when an
event is received, so Telegram can get the events by itself.

This is not like the other push providers where the change information
is processed by PushListenerController directly, but it still saves battery
since the process disconnects from Telegram servers as soon as the
updates are processed.

* [MG] add a longclick listener to UnifiedPush Distributors menu

This prints informations to know if UnifiedPush is working or not.

* [MG] encode the url to pass to p2p.belloworld.it

This is needed if urls includes query (like UP-FCM).

This also means the proxy had to be a little more complex, since nginx didn't do url decoding:
https://gist.github.com/drizzt/93d253bca4f64fae7df2a4544a98c08d

---------

Co-authored-by: Timothy Redaelli <timothy@fsfe.org>